move all extraction code to its own file
